changes made

+++ app/api/auth/delete/[id]/route.ts
@@ -14,14 +14,28 @@ export async function DELETE(request: any, { params }: any) {
   if (user) {
     if (user.id === dataId || user.admin === true) {
       try {
-        const removedUser = await prisma.users.delete({
+        const accountToDelete = await prisma.users.findUnique({
           where: {
             id: dataId,
           },
         });
 
-        if (removedUser)
-          return Response.json({ success: "Account deleted!!", status: 200 });
+        if (accountToDelete?.admin === true && user.id !== accountToDelete.id)
+          return Response.json({
+            error: "You can't delete an admin account",
+            status: 501,
+          });
+
+        if (user.admin === true || accountToDelete?.id === user.id) {
+          const removedUser = await prisma.users.delete({
+            where: {
+              id: dataId,
+            },
+          });
+
+          if (removedUser)
+            return Response.json({ success: "Account deleted!!", status: 200 });
+        }
       } catch (error) {
         console.log(error);
       }
